Basic Genetic Algorithm
One of my first algorithms, programmed in Python to do a genetic-based search, for Basics of AI course in Data Science (MSc).

Why I uploaded it here? To show my friends how it worked and maybe help someone in the future that needs it for some reason.

It is based in some codes I was provided in Java and others that I found in internet that were programmed with Python.