Fuzzy Logic - AC
Basic AC algorithm, programmed in Python to open AC valve (cool down), close valve (warm up) or just keep it in between depending on climate conditions. For Basics of AI course in Data Science (MSc). Valve works with millimetres opened/closed.

Why I uploaded it here? To show my friends how it worked and maybe help someone in the future that needs it for some reason.

It is based in some codes I found on internet and fuzzy data requested.

Rules:

R0: If temperature is fresh or pressure is very low, close valve.
R1: If temperature is room temperature & pressure normal, then keep it like that.
R2:  If temperature is very hot & pressure is strong, open valve.

If something is wrong or can be improved, feel free to do it :)